pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "open-corp"
version = "1.0.0"
description = "AI-powered operations with specialist workers, budget controls, and persistent memory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tinydb>=4.8",
    "httpx>=0.27",
    "click>=8.1",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10,<4.0",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "respx>=0.22",
    "pytest-asyncio>=0.24",
]
training = [
    "yt-dlp",
    "openai-whisper",
    "pypdf>=4.0",
    "html2text>=2024.2",
]
broker = ["yfinance>=0.2"]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

[project.scripts]
corp = "scripts.corp:cli"

[tool.setuptools.packages.find]
include = ["framework*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
